Added some Wifi write code - still unsure about how to read

+++ LilFishTanks/LilFishTanks/lil_fish_tanks_arduino/color_sensor.h
@@ -224,7 +224,7 @@ double FindMatch()
 	}
 }
 
-double ScanColor()
+long ScanColor()
 {
 	// Scan for color
 	CS.read();
